Add branching point for NSFW channels
